refactor(config-ui): move connection-status to component from store
