+++ config-ui/src/store/connections/index.ts
@@ -18,4 +18,3 @@
 
 export * from './types';
 export * from './context';
-export * from './status';
